rbac guard added

+++ vat-calculator-backend/src/account/account.controller.ts
@@ -8,16 +8,23 @@ import {
   Param,
   Put,
   Query,
+  UnprocessableEntityException,
+  UseGuards,
 } from '@nestjs/common';
 import { AccountService } from './account.service';
 import { ApiPaginatedResponse } from 'src/prisma/decorator';
 import { FetchAccountsDto, updateAccountDto } from './dto';
+import { JwtGuard, RolesGuard } from 'src/auth/guard';
+import { HasRoles } from 'src/auth/decorator/has-roles.decorator';
+import { roles } from 'src/auth/dto/enums';
 
+@UseGuards(JwtGuard, RolesGuard)
 @Controller('account')
 export class AccountController {
   constructor(private accountService: AccountService) {}
 
   @Get()
+  @HasRoles(roles.norAdmin)
   @ApiPaginatedResponse(FetchAccountsDto)
   fetchAccounts(
     @Query('search') search: string = '',
@@ -28,14 +35,20 @@ export class AccountController {
   }
 
   @Put()
+  @HasRoles(roles.norAdmin)
   @HttpCode(HttpStatus.ACCEPTED)
   updateAccountInfo(@Body() userInfo: updateAccountDto) {
     return this.accountService.editAccount(userInfo);
   }
 
   @Delete(':accountId')
+  @HasRoles(roles.norAdmin)
   @HttpCode(HttpStatus.ACCEPTED)
   deleteAccount(@Param('accountId') accountId: string) {
+    const accountIdNumber = parseInt(accountId);
+    if (Number.isNaN(accountIdNumber)) {
+      throw new UnprocessableEntityException('The account ID is not a number');
+    }
     return this.accountService.deleteAccount(parseInt(accountId));
   }
 }
